+++ api/src/formatters/properties.ts
@@ -1,5 +1,6 @@
 import * as properties from 'properties';
 import { Exporter, IntermediateTranslationFormat, Parser } from '../domain/formatters';
+import * as iconv from 'iconv-lite';
 
 export const propertiesParser: Parser = async (data: string) => {
   const parsed = await new Promise((resolve, reject) => {
@@ -36,5 +37,6 @@ export const propertiesParser: Parser = async (data: string) => {
 
 export const propertiesExporter: Exporter = async (data: IntermediateTranslationFormat) => {
   const out = data.translations.reduce((acc, obj) => ({ ...acc, [obj.term]: obj.translation }), {});
-  return properties.stringify(out);
+  const result = properties.stringify(out);
+  return iconv.encode(result, 'ISO-8859-1');
 };